[SYCL] Move reduce/scan functors to functional.hpp
Avoid duplicated code in sub_group.hpp and sub_group_host.hpp.
